+++ src/gestures/Shared.hpp
@@ -1,0 +1,4 @@
+#include <cstdint>
+
+// can be one of @eTouchGestureDirection or a combination of them
+using GestureDirection = uint32_t;
